backend/requirements.txt
fastapi[standard]
pydantic-settings
sqlalchemy[asyncio]
asyncpg
aiofiles
pypdf
pytest
